Continue rather than returning in the AutoService loop.
The old code would return as soon as it found a `META-INF/services` file that contributed nothing to the list of services to be generated. It should have continued instead, to go on to the next provider interface.

Fixes https://github.com/google/auto/issues/839. Thanks to @bbyk for the bug report.

RELNOTES=Fixed a bug in AutoServiceProcessor that could lead to some services not being processed.
PiperOrigin-RevId: 375565534